Exercise 4.21

+++ blog_list/controllers/blogs.js
@@ -2,13 +2,6 @@ const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
 const jwt = require('jsonwebtoken')
-const getTokenFrom = request => {
-    const authorization = request.get('authorization')
-    if (authorization && authorization.startsWith('Bearer ')) {
-        return authorization.replace('Bearer ', '')
-    }
-    return null
-}
 
 blogsRouter.get('/', async (request, response) => {
     const blogs = await Blog.find({}).populate('creator')
@@ -38,8 +31,19 @@ blogsRouter.post('/', async (request, response, next) => {
 blogsRouter.delete('/:id', async (request, response, next) => {
     const id = request.params.id
     try {
-        await Blog.findByIdAndDelete(id)
-        response.status(204).end()
+        const decodedToken = jwt.verify(request.token, process.env.SECRET)
+        if (!decodedToken.id) {
+            return response.status(401).json({ error: 'token invalid' })
+        }
+        const user = await User.findById(decodedToken.id)
+        const blog = await Blog.findById(id)
+        
+        if (blog.creator.toString() === user._id.toString()) {
+            await Blog.findByIdAndDelete(id)
+            response.status(204).end()
+        } else {
+            return response.status(403).json({ error: 'incorrect user' })
+        }
     } catch(exception) {
         next(exception)
     }
